+++ R/get_data.R
@@ -134,13 +134,45 @@ get_timeseries_full_data <- function() {
   data
 }
 
+get_datahub_fix_ch <- function(country = NULL, startdate = "2020-01-22", lev = 1, verbose = FALSE) {
+
+  orig_data <-
+    get_datahub(country = country, startdate = startdate, lev = lev, verbose = verbose)
+
+  orig_data_ch_2 <-
+    get_datahub(country = "Switzerland", startdate = startdate, lev = 2, verbose = verbose)
+
+  message("replace hosp data (",paste(.hosp_vars, collapse = ","), ") in lev1 dataset with lev2 swiss data")
+
+  # aggregate hosp data at country level
+  orig_data_ch_1 = orig_data_ch_2 %>%
+    select(date, all_of(as.vector(.hosp_vars))) %>%
+    group_by(date) %>%
+    summarise_if(is.numeric, sum, na.rm = TRUE) %>% # todo, use only hospvars
+    ungroup() %>%
+    mutate(Country.Region = "Switzerland")
+  orig_data_ch_1 = orig_data_ch_1[, c("Country.Region", setdiff(names(orig_data_ch_1),"Country.Region"))]
+
+  if (!identical(orig_data_ch_1$date, orig_data$date[orig_data$Country.Region == "Switzerland"]))
+    warning("Not same dates in lev1 and lev2 for CH")
+
+  commondates = match(orig_data_ch_1$date ,orig_data$date[orig_data$Country.Region == "Switzerland"])
+
+  orig_data[orig_data$Country.Region == "Switzerland" & (orig_data$date %in% orig_data_ch_1$date), .hosp_vars] <-
+    orig_data_ch_1[(orig_data_ch_1$date %in% orig_data$date) , .hosp_vars]
+
+  list(orig_data = orig_data, orig_data_ch_2 = orig_data_ch_2)
+}
+
+
 #' Get timeseries full data from datahub
 #' @rdname get_datahub
 #'
 #' @param country character country, to chose with lev = 2
-#' @param stardate character staring date
+#' @param startdate character staring date
 #' @param lev integer 1 for country level, 2 for reagions
 #' @param verbose logical. Print data sources? Default FALSE (opposite from \code{covid19})
+#' @param hosp logical. If TRUE hospitalised detailed data are retrieved. Default TRUE since release 2.3.1
 #'
 #' @details data sourced from https://github.com/covid19datahub/COVID19/
 #'
@@ -150,8 +182,8 @@ get_timeseries_full_data <- function() {
 #' @import dplyr
 #'
 #' @export
-get_datahub = function(country = NULL, stardate = "2020-01-22", lev = 1, verbose = FALSE) {
-  message("get_datahub: country = ", country, "/ startdate = ", stardate, "/ level = ", lev)
+get_datahub = function(country = NULL, startdate = "2020-01-22", lev = 1, verbose = FALSE, hosp = TRUE) {
+  message("get_datahub: country = ", country, "/ startdate = ", startdate, "/ level = ", lev)
   if (!is.null(country)) {
     # remap country )
     country = recode(country,
@@ -165,22 +197,24 @@ get_datahub = function(country = NULL, stardate = "2020-01-22", lev = 1, verbose
                      "U.S. Virgin Islands" = "U.S. Virgin Islands",
                      )
   }
-  dataHub <- covid19(country = country, start = stardate, level = lev, verbose = verbose) # select level2 to add states
+  dataHub <- covid19(country = country, start = startdate, level = lev, verbose = verbose) # select level2 to add states
 
+  vars = c("date", "tests", "confirmed", "recovered", "deaths", "hosp", "population")
+  if (hosp) {
+    vars = append(vars, setdiff(.hosp_vars_datahub, vars), which(vars == "hosp"))
+  }
   # if Hong Kong was chosen in country
   if ((is.null(dataHub) || nrow(dataHub) == 0) && lev == 1 && country == "Hong Kong") {
     message("Taking Hong Kong from chinese data from level 2")
-    dataHub <- covid19("China",2, start = stardate, verbose = verbose) %>% ungroup() %>%
-      select( administrative_area_level_2, date, tests,
-             confirmed, recovered, deaths, hosp, population) %>%
+    dataHub <- covid19("China",2, start = startdate, verbose = verbose) %>% ungroup() %>%
+      select( administrative_area_level_2, !!vars) %>%
       rename(Country.Region = administrative_area_level_2) %>%
       filter(Country.Region == "Hong Kong")
   } else if (!is.null(dataHub) && nrow(dataHub) > 0) {
     adminvar = paste("administrative_area_level", lev, sep = "_")
     # select varaibles for backwards compatibility + some additional variables
     dataHub = dataHub %>% ungroup() %>% select(!!adminvar, #id, # at the moment removing ID
-                                               date, tests,
-                                               confirmed, recovered, deaths, hosp, population) %>%
+                                               !!vars) %>%
       rename(Country.Region = !!adminvar) # rename country variable
 
     if (lev == 1) {
@@ -201,12 +235,11 @@ get_datahub = function(country = NULL, stardate = "2020-01-22", lev = 1, verbose
     # "Virgin Islands, U.S." belongs to USA
 
     # if either china or hong kong missing
-    if (length(setdiff(c("Hong Kong","China"), dataHub$Country.Region))>0) {
+    if (length(setdiff(c("Hong Kong","China"), dataHub$Country.Region))==1) {
       if (lev == 1) {
         message("Taking chinese data from level 2")
-        dataMiss <- covid19("China",2, start = stardate, verbose = verbose) %>% ungroup() %>%
-          select(administrative_area_level_1, administrative_area_level_2, date, tests,
-                 confirmed, recovered, deaths, hosp, population) %>%
+        dataMiss <- covid19("China",2, start = startdate, verbose = verbose) %>% ungroup() %>%
+          select(administrative_area_level_1, administrative_area_level_2, !!vars) %>%
           rename(Country.Region = administrative_area_level_1)
         # separate Hong Kong
         if (is.null(country) && (!("Hong Kong" %in% dataHub$Country.Region))) {
@@ -232,12 +265,14 @@ get_datahub = function(country = NULL, stardate = "2020-01-22", lev = 1, verbose
   }
   if (!is.null(dataHub) && nrow(dataHub) > 0) {
     # adjust recovered where they do not make sense, e.g. France lev 2
-    dataHub$recovered = pmin(dataHub$recovered, dataHub$confirmed)
-    dataHub$deaths    = pmin(dataHub$deaths, dataHub$confirmed)
+    dataHub$recovered[is.na(dataHub$recovered)] = 0 #TODO: review, required only with COVID19 2.3.1
+    dataHub$recovered = pmin(dataHub$recovered, dataHub$confirmed, na.rm = TRUE)
+    dataHub$deaths    = pmin(dataHub$deaths, dataHub$confirmed, na.rm = TRUE)
 
     # compute active
     dataHub = dataHub %>%
-      mutate(active = confirmed - deaths - recovered)
+      mutate(active = confirmed - deaths - recovered) %>%
+      mutate(active = replace_na(active, 0)) # TODO review
 
     # convert integers into numeric
     dataHub[,sapply(dataHub, class) == "integer"] = dataHub[,sapply(dataHub, class) == "integer"] %>% sapply(as.numeric)
@@ -251,8 +286,20 @@ get_datahub = function(country = NULL, stardate = "2020-01-22", lev = 1, verbose
     #TODO: arrange should go descending, many rows could be filtered out for many countries#
     dataHub = dataHub %>% filter(date <= maxdate) %>% arrange(Country.Region, date)
 
+    if (hosp) {
+      message("Edit hosp data")
+      # hospitalised must be
+      dataHub$icuvent = replace_na(dataHub$vent,0)+replace_na(dataHub$icu,0) # check NAs
+      dataHub$icuvent[is.na(dataHub$vent) & is.na(dataHub$icu) ] = NA # NA if both were Na
+      dataHub$hosp = pmax(dataHub$hosp, dataHub$icuvent) # generally it can be lower
+      dataHub = dataHub %>% select(-icu,-vent) # remove icu and vent
+      if (any(dataHub$active < dataHub$hosp)) {
+        warning(sum(dataHub$active < dataHub$hosp), " cases with active < hosp")
+      }
+    }
+
   }  else {
-    warning("Data not found for country = ", country, " startdate = ", stardate, " level = ", lev)
+    warning("Data not found for country = ", country, " startdate = ", startdate, " level = ", lev)
   }
 
   dataHub
@@ -275,6 +322,12 @@ get_timeseries_by_contagion_day_data <- function(data) {
   if (!("tests" %in% names(data))) {
     data$tests = data$hosp = rep(0,nrow(data))
   }
+  # Select column names for which new_vars to be created
+  new_vars = intersect(setdiff(get_aggrvars(), "population"), names(data))
+
+  new_calc = function(x) {
+    x - replace_na(lag(x),0)
+  }
   data1 <- data %>%
     arrange(desc(Country.Region), date) %>%
     mutate(no_contagion = case_when(
@@ -291,12 +344,23 @@ get_timeseries_by_contagion_day_data <- function(data) {
     )) %>%
     select(-incremental, -offset, -no_contagion, -tmp) %>%
     # added replace_na around lag to avoid NA in fist position
-    mutate(new_confirmed = confirmed - replace_na(lag(confirmed),0),
-        new_deaths = deaths - replace_na(lag(deaths),0),
-        new_active = active - replace_na(lag(active),0),
-        new_recovered = recovered - replace_na(lag(recovered),0),
-        new_tests = tests - replace_na(lag(tests),0),
-        new_hosp = hosp - replace_na(lag(hosp),0)) %>%
+    mutate(
+      across(all_of(new_vars), new_calc, .names="new_{col}") # use all_of
+    ) %>%
+    # mutate(
+    #   tot_hosp = cumsum(new_hosp, na.rm = TRUE),
+    #   tot_icuvent = cumsum(new_icuvent, na.rm = TRUE),
+    #   )
+    # added replace_na around lag to avoid NA in fist position
+    # mutate(new_confirmed = confirmed - replace_na(lag(confirmed),0),
+    #     new_deaths = deaths - replace_na(lag(deaths),0),
+    #     new_active = active - replace_na(lag(active),0),
+    #     new_recovered = recovered - replace_na(lag(recovered),0),
+    #     new_tests = tests - replace_na(lag(tests),0),
+    #     new_hosp = hosp - replace_na(lag(hosp),0)
+    #     # new_vent = vent - replace_na(lag(vent),0),
+    #     # new_icu = icu - replace_na(lag(icu),0)
+    #     ) %>%
     # mutate(new_confirmed = if_else(is.na(new_confirmed), 0, new_confirmed)) %>%
     # mutate(new_deaths = if_else(is.na(new_deaths), 0, new_deaths)) %>%
     # mutate(new_active = if_else(is.na(new_active), 0, new_active)) %>%
@@ -319,13 +383,13 @@ get_timeseries_by_contagion_day_data <- function(data) {
 #'
 #' @export
 get_timeseries_global_data <- function(data, new = FALSE){
-  vars = get_aggrvars()
+  vars =  intersect(get_aggrvars(), names(data))
   if (!new)
     vars = setdiff(vars, grep("new",vars, value = TRUE))
   vars = intersect(vars, names(data))
   data %>%
     group_by(date) %>%
-    summarize_at(vars, sum) %>%
+    summarize_at(vars, sum, na.rm = TRUE) %>%
     ungroup()
 }