+++ R/mod_individual_country.R
@@ -14,10 +14,12 @@ mod_ind_country_ui <- function(id){
     hr(),
     tags$head(tags$style(HTML(".small-box {width: 300px; margin: 20px;}"))),
     mod_caseBoxes_ui(ns("ind_count-boxes")),
+    mod_caseBoxes_ui(ns("ind_count-boxes_hosp"), hosp = TRUE),
     hr(),
     div(
       uiOutput(ns("ind_from_nth_case"))
     ),
+    hr(),
     fluidRow(
      # withSpinner(uiOutput(ns("ind_barplots")))
       withSpinner(mod_bar_plot_day_contagion_ui(ns("ind_bar_plot_day_contagion")))
@@ -27,13 +29,23 @@ mod_ind_country_ui <- function(id){
     fluidRow(
       column(6,
              br(),
-             div( h4("Total cases"), align = "center",
+             div( h4("Covid-19 time evolution"), align = "center",
              div(style = "visibility: hidden", radioButtons("dummy1", "", choices = "dummy")),
              withSpinner(mod_plot_log_linear_ui(ns("ind_plot_area_tot"), area = TRUE))
              )
       ),
       column(6,
-             withSpinner(mod_compare_nth_cases_plot_ui(ns("ind_lines_points_plots_tot"), tests = TRUE, hosp = TRUE))
+             br(),
+             div( h4("Time evolution of Hospitalised cases"), align = "center",
+                  div(style = "visibility: hidden", radioButtons("dummy1", "", choices = "dummy")),
+                  withSpinner(mod_plot_log_linear_ui(ns("ind_plot_areahosp_tot"), area = TRUE))
+             )
+      )
+    ),
+    fluidRow(
+
+      column(6,
+             withSpinner(mod_compare_nth_cases_plot_ui(ns("ind_lines_points_plots_tot"), tests = TRUE, hosp = TRUE, oneMpop = FALSE))
       )
     ),
     # hr(),
@@ -89,13 +101,22 @@ areamapUI = function(id, country){
           column(6,
                  withSpinner(uiOutput(ns("map_countries_death")))
           )
+        ),
+        fluidRow(
+          column(6,
+                 withSpinner(uiOutput(ns("map_countries_hosp")))
+          ),
+          column(6,
+                 withSpinner(uiOutput(ns("map_countries_hosp_1M_pop")))
+          )
         )
     )
   )
 }
 #' ind_country Server Function
 #'
-#' @param data data.frame
+#' @param data data.frame with Country Level data
+#' @param data2 data.frame with Country Level 2 data
 #' @param countries reactive character vector
 #' @param nn min number of cases for used to filter country data
 #' @param w number of days of outbreak. Default 7
@@ -105,14 +126,17 @@ areamapUI = function(id, country){
 #' @import shiny
 #'
 #' @noRd
-mod_ind_country_server <- function(input, output, session, data, country , nn = 1, w = 7){
+mod_ind_country_server <- function(input, output, session, data, data2, country , nn = 1, w = 7){
   ns <- session$ns
 
   message("mod_ind_country_server")
   output$ind_from_nth_case<- renderUI({
     HTML(paste(
-         paste0("Some Cantons are not providing Recovered data."),
-         paste0("1st day is the day when ", nn ," confirmed cases are reached."), sep = "<br/>"))
+         #paste0("Some Cantons are not providing Recovered data. Test "),
+         message_missing_data("Recovered and Tests",where = "most of Cantons"),
+         #paste0("1st day is the day when ", nn ," confirmed cases are reached."),
+         message_firstday(nn),
+         message_hosp_data(where = "Cantons"), sep = "<br/>"))
   })
 
   message("process individual country ", country)
@@ -122,31 +146,48 @@ mod_ind_country_server <- function(input, output, session, data, country , nn = 
       filter(contagion_day > 0) %>%
       arrange(desc(date))
 
+  lw_country_data = lw_vars_calc(country_data)
+
+  # create datasets for box merging today with data7
+  lw_country_data_today = country_data %>% filter(date == max(date)) %>%
+    left_join(lw_country_data %>% select(-population))
+
   country_data_today <- country_data %>%
       filter(date == max(date))
 
   # Boxes ----
   callModule(mod_caseBoxes_server, "ind_count-boxes", country_data_today)
 
-  # tables ----
-  callModule(mod_add_table_server, "ind_add_table_country", country_data,  maxrowsperpage = 10)
-  # plots ----
-  levs <- sort_type_hardcoded()
-  country_data_area = country_data
-  if (sum(country_data$hosp)>0) {
-    message("Adding hospitalised data for ", country)
-    levs = c(levs, "hosp")
-    country_data_area$active = country_data_area$active - country_data_area$hosp
-  }
-  message("n for ", country, " = ", nn)
+  # Boxes ----
+  callModule(mod_caseBoxes_server, "ind_count-boxes_hosp", lw_country_data_today, hosp = TRUE)
 
 
   callModule(mod_bar_plot_day_contagion_server, "ind_bar_plot_day_contagion", country_data, nn = nn)
 
-  # for country plot start from the beginning
-  df_tot = tsdata_areplot(country_data_area,levs, nn = nn) # start from day with >nn
+  # tables ----
+  callModule(mod_add_table_server, "ind_add_table_country", country_data,  maxrowsperpage = 10)
+  # plots ----
+  levs <- areaplot_vars()
+  country_data_area = country_data
+  active_hosp = FALSE
+  if (sum(country_data$hosp, na.rm = TRUE)>0) {
+    message("Adding hospitalised data for areaplot for ", country)
+    levs = c(levs, "hosp")
+    active_hosp = TRUE
+  }
+  message("n for ", country, " = ", nn)
 
-  callModule(mod_plot_log_linear_server, "ind_plot_area_tot", df = df_tot, type = "area")
+  # for country plot start from the beginning
+  df_tot = tsdata_areplot(country_data_area, levs, nn = nn) # start from day with >nn
+
+  callModule(mod_plot_log_linear_server, "ind_plot_area_tot", df = df_tot, type = "area", active_hosp = active_hosp)
+
+  # for country plot start from the beginning
+  levs <- areaplot_hospvars()
+
+  df_hosp = tsdata_areplot(country_data_area, levs, nn = 1) # start from day with >nn
+
+  callModule(mod_plot_log_linear_server, "ind_plot_areahosp_tot", df = df_hosp, type = "area", hosp = TRUE)
 
   callModule(mod_compare_nth_cases_plot_server, "ind_lines_points_plots_tot", country_data , nn = nn, w = w, istop = FALSE)
 
@@ -154,10 +195,15 @@ mod_ind_country_server <- function(input, output, session, data, country , nn = 
 # # ##### country split within areas #############################################
 
 #   # Data ----
-  area_data_2 = get_datahub(country = country, lev = 2, verbose = FALSE)
+  if (missing(data2)) {
+    area_data_2 = get_datahub(country = country, lev = 2, verbose = FALSE)
 
-  area_data_2 = area_data_2 %>%
-    get_timeseries_by_contagion_day_data()
+    area_data_2 = area_data_2 %>%
+      get_timeseries_by_contagion_day_data()
+
+  } else
+    area_data_2 = data2
+
 
   area_data_2_aggregate <-
     build_data_aggr(area_data_2)
@@ -277,6 +323,19 @@ mod_country_area_maps_server <- function(input, output, session, data, country){
   callModule(mod_map_area_calc_server, "map_ind_death", df = data_maps,  area2_map,
              area = country, variable = "death", max.pop = 0, countrymap = TRUE)
 
+  #maps hosp
+  output[["map_countries_hosp"]] <- renderUI({
+    mod_map_area_calc_ui(ns("map_ind_hosp"))
+  })
+  callModule(mod_map_area_calc_server, "map_ind_hosp", df = data_maps,  area2_map,
+             area = country, variable = "hospitalised", max.pop = 0, countrymap = TRUE)
+
+  #maps hosp per population
+  output[["map_countries_hosp_1M_pop"]] <- renderUI({
+    mod_map_area_calc_ui(ns("map_ind_hosp_1M_pop"))
+  })
+  callModule(mod_map_area_calc_server, "map_ind_hosp_1M_pop", df = data_maps,  area2_map,
+             area = country, variable = "hospitalised over 1M", max.pop = 0, countrymap = TRUE)
 }