+++ R/mod_lineplots_day_contagion.R
@@ -39,7 +39,7 @@ mod_lineplots_day_contagion_server <- function(input, output, session, countries
   #     select(Country.Region) %>%
   #     pull()
   # })
-  mindate = min(countries_data$date[countries_data$confirmed>nn])
+  mindate = min(countries_data$date[countries_data$confirmed>nn], na.rm = TRUE)
   countries_data = countries_data %>% filter(date > mindate)
 
   statuses <- c("confirmed", "deaths", "recovered", "active")