+++ man-roxygen/ex-plot_log_linear.R
@@ -4,7 +4,7 @@ if (interactive()) {
   library(ggplot2)
   library(tidyr)
   #sapply(file.path("R",list.files("R")), source)
-  #devtools::load_all()
+  devtools::load_all()
   long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
   ui <- fluidPage(
     Covid19Mirai:::mod_plot_log_linear_ui("test", select = FALSE, area = FALSE)
@@ -33,7 +33,7 @@ if (interactive()) {
         get_timeseries_global_data()
     })
 
-    levs <- sort_type_hardcoded()
+    levs <- areaplot_vars()
 
     df_country = reactive({
       tsdata_areplot(country(),levs)
@@ -78,7 +78,7 @@ if (interactive()) {
         get_timeseries_global_data()
     })
 
-    levs <- sort_type_hardcoded()
+    levs <- areaplot_vars()
 
     df_country = reactive({
       tsdata_areplot(country(),levs)
@@ -117,7 +117,7 @@ if (interactive()) {
         filter(contagion_day > 0) %>%
         arrange(desc(date))
 
-    levs <- sort_type_hardcoded()
+    levs <- areaplot_vars()
 
     df_area = purrr::map(unique(country_data$Country.Region),
                            function(un) {
@@ -133,6 +133,56 @@ if (interactive()) {
         distinct() #%>% .$Country.Region
     })
     callModule(mod_plot_log_linear_server,"test", df = df_country, type = "area", countries = countries)
+  }
+  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
+}
+
+
+# with select and hospedalised data
+if (interactive()) {
+  library(shiny)
+  library(dplyr)
+  library(ggplot2)
+  library(tidyr)
+  sapply(file.path("R",list.files("R")), source)
+  #devtools::load_all()
+  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
+  ui <- fluidPage(
+    Covid19Mirai:::mod_plot_log_linear_ui("test", select = TRUE, area = TRUE)
+  )
+  server <- function(input, output) {
+
+    orig_data_ch2 <- get_datahub(country = "Switzerland", lev = 2) %>%
+      get_timeseries_by_contagion_day_data()
+
+    n = 1
+    w = 1
+    data_filtered <-
+      orig_data_ch2 %>%
+      Covid19Mirai:::rescale_df_contagion(n = n, w = w)
+
+
+    country_data <- data_filtered %>%
+      #filter(Country.Region %in% "Switzerland") %>%
+      filter(contagion_day > 0) %>%
+      arrange(desc(date))
+
+    levs <- areaplot_hospvars()
+
+    df_area = purrr::map(unique(country_data$Country.Region),
+                         function(un) {
+                           dat = tsdata_areplot(country_data[country_data$Country.Region == un, ], levs, nn = 10) #n = 0 for area plot
+                           dat$Country.Region = rep(un, nrow(dat))
+                           dat
+                         })
+    df_country = Reduce("rbind",df_area)
+
+    countries <- reactive({
+      country_data %>%
+        select(Country.Region) %>%
+        distinct() #%>% .$Country.Region
+    })
+    callModule(mod_plot_log_linear_server,"test", df = df_country, type = "area", countries = countries, hosp = TRUE)
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }